moved the toggle from the top into the SearchResultsTable, and made the state save locally in each row
